3rdparty: Update libjpeg-turbo to 2.0.6, libtiff to 4.2.0, glew to 2.2.0 and Qt to 5.15.2
Add sanity checks for Qt because its build system requires either Perl or
PowerShell >= 5.0 for generating qmimeprovider_database.cpp. If Perl is missing,
there is a fallback to a PowerShell command. But if the installed PowerShell
version is too old (happened on my Windows 7 system which had PowerShell 2.0),
it can cause an issue with generating qmimeprovider_database.cpp, resulting in
the following error:
  build\qt5\src\qt5\qtbase\src\corelib\.rcc\debug\qmimeprovider_database.cpp(142529): error C2059: syntax error: ';' (compiling source file mimetypes\qmimeprovider.cpp)